refactor(config): rename LambdaExtension -> LambdaConfig, inline into mod.rs
`LambdaExtension` collides nominally with "the Datadog Lambda Extension"
(this entire project). `LambdaConfig` reads more naturally for the
extension type that holds Lambda-specific configuration fields.

While here, fold the standalone lambda_extension.rs into config/mod.rs
so consumers don't need a separate module hop. The inlined code uses
fully-qualified `datadog_agent_config::merge_fields!` /
`datadog_agent_config::merge_string!` invocations to coexist with the
legacy `#[macro_export]` macros still at the top of mod.rs — both go
away together once the migration onto the upstream Config lands.

Renames carried through:
  LambdaExtension  -> LambdaConfig
  LambdaSource     -> LambdaConfigSource
  mod lambda_extension -> (gone; inlined)

All 33 LambdaConfig tests still pass.
